+++ src/components/Fields/AssociatedClassification.tsx
@@ -1,11 +1,11 @@
-import React, { useState, useEffect } from "react";
+import React, { useEffect } from "react";
 import { Box, Flex, FormLabel, Spinner, Text } from "@chakra-ui/react";
 import { v4 as uuidv4 } from "uuid";
-import { useField } from "formik";
 import { useAppSelector } from "redux/hooks";
 import { Maybe } from "api/graphql/types.generated";
 import { selectors } from "redux/slices/application";
 import { useMediaQueries } from "utils/hooks/mediaqueries";
+import { useAssociatedLogic } from "./hooks";
 
 interface Props {
   label: string;
@@ -14,11 +14,6 @@ interface Props {
   isCollapsed?: boolean;
   factors: Factor[];
   maxLoop: Maybe<string> | undefined;
-}
-
-interface State {
-  variations: number[][][];
-  isMounted: boolean;
 }
 
 interface Factor {
@@ -48,7 +43,7 @@ export const AssociatedClassification: React.FC<Props> = ({
     totalClick,
     maxVariations,
     isFinished,
-  } = useAssociatedLogic(factors, name, maxLoop);
+  } = useAssociatedLogic(factors, name, maxLoop, TOTAL_CARDS);
 
   const drawerIsOpen = useAppSelector(selectors.drawerIsOpen);
 
@@ -133,8 +128,8 @@ export const AssociatedClassification: React.FC<Props> = ({
       {maxLoop && maxVariations >= 1 && (
         <Text mt="15px" fontSize="xs">
           {maxVariations > parseInt(maxLoop)
-            ? `${totalClick} / ${parseInt(maxLoop) - 1}`
-            : `${totalClick}  / ${Math.max(maxVariations) - 1}`}
+            ? `${totalClick} / ${parseInt(maxLoop)}`
+            : `${totalClick}  / ${Math.max(maxVariations)}`}
         </Text>
       )}
       {!isCollapsed && (
@@ -158,134 +153,4 @@ export const AssociatedClassification: React.FC<Props> = ({
       )}
     </Box>
   );
-};
-
-// eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
-export const useAssociatedLogic = (
-  factors: Factor[],
-  name: string,
-  maxLoop: Maybe<string>
-) => {
-  const [field, , helpers] = useField(name);
-
-  const [state, setState] = useState<State>({
-    variations: [],
-    isMounted: false,
-  });
-  const [totalClick, setClick] = useState(0);
-  const filteredFactors = factors?.filter((f) => f !== null);
-  const modalitiesPerFactor = filteredFactors
-    ?.map((f) => f.modalities?.length)
-    .filter((m) => m !== 0);
-  const totalVariations = modalitiesPerFactor?.reduce((a, b) => a * b, 1);
-
-  const getMaxVariation: any = (n: number, k: number) => {
-    const factorialize: any = (num: number) => {
-      if (num < 0) return -1;
-      else if (num === 0) return 1;
-      else {
-        return num * factorialize(num - 1);
-      }
-    };
-
-    const _A = (n: number, k: number) => {
-      return factorialize(n) / factorialize(n - k);
-    };
-
-    return _A(n, k) / factorialize(k);
-  };
-
-  const maxVariations = React.useMemo(() => {
-    if (totalVariations) return getMaxVariation(totalVariations, TOTAL_CARDS);
-  }, [totalVariations]);
-
-  const generate = () => {
-    if (maxVariations - 1 === state.variations.length) {
-      console.log("End of variations");
-      return;
-    }
-    if (!modalitiesPerFactor) {
-      return;
-    }
-    const randomize = () => {
-      return modalitiesPerFactor?.map((m) => Math.floor(Math.random() * m));
-    };
-
-    const card1 = randomize();
-    const card2 = randomize();
-
-    const variation = [card1, card2];
-
-    const cardsAreSame = (arrA: number[], arrB: number[]) => {
-      return JSON.stringify(arrA) === JSON.stringify(arrB);
-    };
-
-    if (cardsAreSame(card1, card2)) {
-      console.log("same cards");
-      generate();
-    } else if (
-      state.variations.some(
-        (v) => JSON.stringify(v) === JSON.stringify(variation)
-      ) ||
-      state.variations.some(
-        (v) => JSON.stringify(v) === JSON.stringify(variation.reverse())
-      )
-    ) {
-      console.log("Variation already exists");
-      generate();
-    } else {
-      setState({
-        ...state,
-        variations: [...state.variations, variation],
-        isMounted: true,
-      });
-    }
-  };
-  const handleClick = (cardIdx: number) => {
-    generate();
-    setClick(totalClick + 1);
-
-    const formatPayload = () => {
-      const lastVariation = state.variations[state.variations.length - 1];
-
-      const format = (el: number) => {
-        return filteredFactors?.map((f, idx) => {
-          return {
-            [f.title]: f.modalities[lastVariation[el][idx]].description,
-          };
-        });
-      };
-
-      // TODO: replace format(0) by dynamic value (come from TOTAL_CARDS)
-      return {
-        variations: [format(0), format(1)],
-        choice: cardIdx,
-      };
-    };
-
-    if (!field.value) {
-      helpers.setValue([formatPayload()]);
-    } else {
-      helpers.setValue([...field.value, formatPayload()]);
-    }
-  };
-
-  // TODO: refactor this
-  const isFinished =
-    totalClick + 1 ===
-      (maxVariations - 1 > (typeof maxLoop === "string" && parseInt(maxLoop))
-        ? maxLoop && parseInt(maxLoop)
-        : maxVariations) ||
-    field.value?.length ===
-      ((maxLoop && parseInt(maxLoop) - 1) || maxVariations);
-  return {
-    generate,
-    handleClick,
-    setState,
-    state,
-    filteredFactors,
-    totalClick,
-    maxVariations,
-    isFinished,
-  };
 };
