+++ src/components/Fields/MonoThumbnail.tsx
@@ -14,7 +14,8 @@ interface Factor {
 }
 import { selectors } from "redux/slices/application";
 import { RenderInput } from "components/CreateSurvey/CreateForm/InputsPreview/Card/utils";
-import { useAssociatedLogic } from "./AssociatedClassification";
+import { useAssociatedLogic } from "./hooks";
+import { TitleDivider } from "components/TitleDivider";
 
 interface Props {
   label: string;
@@ -44,8 +45,8 @@ export const MonoThumbnail: React.FC<Props> = ({
     filteredFactors,
     totalClick,
     maxVariations,
-    isFinished,
-  } = useAssociatedLogic(factors, name, maxLoop);
+    checkIsFinished,
+  } = useAssociatedLogic(factors, name, maxLoop, TOTAL_CARDS);
   const drawerIsOpen = useAppSelector(selectors.drawerIsOpen);
 
   const Card = ({ index }: { index: number }) => {
@@ -101,7 +102,7 @@ export const MonoThumbnail: React.FC<Props> = ({
     !drawerIsOpen && generate();
   }, [drawerIsOpen]);
 
-  if (isFinished) {
+  if (checkIsFinished() && !drawerIsOpen) {
     return (
       <Text variant="smallTitle">
         Nous avons bien pris en compte votre sélection !
@@ -115,8 +116,8 @@ export const MonoThumbnail: React.FC<Props> = ({
       {maxLoop && maxVariations >= 1 && (
         <Text mt="15px" fontSize="xs">
           {maxVariations > parseInt(maxLoop)
-            ? `${totalClick} / ${parseInt(maxLoop) - 1}`
-            : `${totalClick}  / ${Math.max(maxVariations) - 1}`}
+            ? `${totalClick} / ${parseInt(maxLoop)}`
+            : `${totalClick}  / ${Math.max(maxVariations)}`}
         </Text>
       )}
       {!isCollapsed && (
@@ -130,26 +131,39 @@ export const MonoThumbnail: React.FC<Props> = ({
             <Text mt="15px" fontSize="xs">
               {helpText}
             </Text>
-            <Box mt="10">
+            <TitleDivider title="" />
+            <Box>
               <Formik
                 initialValues={{ ...mono_thumbnail_input }}
                 onSubmit={() => console.log("")}
               >
-                <Form>
-                  {mono_thumbnail_input && (
-                    <RenderInput input={mono_thumbnail_input} />
-                  )}
-                </Form>
+                {({ values }) => {
+                  return (
+                    <Form>
+                      {mono_thumbnail_input && (
+                        <Box p="0 5%">
+                          <RenderInput input={mono_thumbnail_input} />
+                        </Box>
+                      )}
+                      <Box
+                        d="flex"
+                        justifyContent="flex-end"
+                        w="100%"
+                        mt="20px"
+                        pr="30px"
+                      >
+                        <Button
+                          type="button"
+                          variant="rounded"
+                          onClick={() => handleClick(0, values)}
+                        >
+                          Valider ma réponse
+                        </Button>
+                      </Box>
+                    </Form>
+                  );
+                }}
               </Formik>
-              <Box d="flex" justifyContent="flex-end" w="100%">
-                <Button
-                  type="button"
-                  variant="rounded"
-                  onClick={() => handleClick(0)}
-                >
-                  Valider ma réponse
-                </Button>
-              </Box>
             </Box>
           </Box>
         </Flex>
